security/manager/ssl/tests/mochitest/bugs/test_ev_validation_child.html

<html>
<head>
<title></title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<script type="text/javascript">

function finish(state) {
  try {
    window.opener.postMessage(state, "*");
    //the opener should close this window, no
    // need to call simpletest.finish()
    return;
  }
  catch(e){
    if (window.opener) {
      //could not sent the message?
      ok(false,"could no opener window");
    }
  }
}

function onWindowLoad()
{
  if (!window.opener) {
    ok(true, "Direct call, nothing to do");
    SimpleTest.finish();
  }

  var ui = SpecialPowers.wrap(window)
     .QueryInterface(SpecialPowers.Ci.nsIInterfaceRequestor)
     .getInterface(SpecialPowers.Ci.nsIWebNavigation)
     .QueryInterface(SpecialPowers.Ci.nsIDocShell)
     .securityUI;

  var isInsecure = !ui ||
     (ui.state & SpecialPowers.Ci.nsIWebProgressListener.STATE_IS_INSECURE);
  var isBroken = ui &&
     (ui.state & SpecialPowers.Ci.nsIWebProgressListener.STATE_IS_BROKEN);
  var isEV = ui &&
     (ui.state & SpecialPowers.Ci.nsIWebProgressListener.STATE_IDENTITY_EV_TOPLEVEL);
 
  var gotState;
  if (isInsecure)
    gotState = "insecure";
  else if (isBroken)
    gotState = "broken";
  else if (isEV)
    gotState = "EV";
  else
    gotState = "secure";

  finish(gotState);
}

</script>
</head>

<body onload="onWindowLoad()">
 <h1>Security_state_child</h1>
</body>
</html>
